fix: fixed auth reset after reload
